Represented CAMSL as int64 instead of int32

+++ datastore/migrate/data/migrations/1747651104_camsl.up.sql
@@ -1,4 +1,4 @@
-ALTER TABLE geo_point ADD COLUMN camsl INTEGER;
+ALTER TABLE geo_point ADD COLUMN camsl BIGINT;
 
 -- drop UNIQUE constraint of 'point' column
 -- WARNING: we assume that the constraint name is the correct one (it was never explicitly set)
